Refactor Forecast component to use DailyWeatherPanel in Forecast.tsx
