+++ app/components/Forecast.tsx
@@ -16,6 +16,7 @@ import {
   toKph,
 } from "../helper";
 import { BuildPanelProps, ForecastProps } from "../models/props";
+import { DailyWeatherPanel } from "./DailyWeatherPanel";
 
 /**
  * This pagination builds the panels below the bottom element
@@ -26,69 +27,7 @@ import { BuildPanelProps, ForecastProps } from "../models/props";
  * @param {string} props.timezone - The timezone.
  * @returns {JSX.Element} The Forecast component.
  */
-export const Forecast = (props: ForecastProps): JSX.Element => {
-  const { data, timezone } = props;
-
-  /**
-   * This builds the panel with weather information by mapping over the data and pushing its value into elements
-   * @param {Object} value - The weather data object.
-   * @returns {JSX.Element[]} An array of JSX elements representing the weather panel.
-   */
-  const BuildPanel = (value: BuildPanelProps): JSX.Element[] => {
-    const clouds = value.item.clouds + "%";
-    const description = value.item.weather[0].description;
-    const humidity = value.item.humidity + "%";
-    const moon = getMoonPhase(value.item.moon_phase);
-    const sunrise = parseTime(value.item.sunrise, "en-US", timezone);
-    const sunset = parseTime(value.item.sunset, "en-US", timezone);
-    const uvi = value.item.uvi;
-    const rain = value.item.pop * 100 + "%"; //Rain is given to us from 0-1, 1 meaning 100%
-    const dewPoint =
-      toCelsius(value.item.dew_point) + "°C | " + value.item.dew_point + "°F";
-
-    const precipitation =
-      value.item.rain! < 0 || typeof value.item.rain === "undefined"
-        ? "0.00"
-        : value.item.rain +
-          "mm | " +
-          mmToInches(value.item.rain).toFixed(2) +
-          "in";
-
-    const windSpeed =
-      toKph(value.item.wind_speed) + " kph | " + value.item.wind_speed + " mph";
-
-    const windGust =
-      toKph(value.item.wind_gust) + " kph | " + value.item.wind_gust + " mph";
-
-    const collection = [
-      { id: "Forecast", result: description },
-      { id: "Dew Point", result: dewPoint },
-      { id: "Clouds", result: clouds },
-      { id: "Humidity", result: humidity },
-      { id: "Rain", result: rain },
-      { id: "Precipitation", result: precipitation },
-
-      { id: "UV Index", result: uvi },
-      { id: "Moon", result: moon },
-      { id: "Wind", result: windSpeed },
-      { id: "Gust", result: windGust },
-      { id: "Sunrise", result: sunrise },
-      { id: "Sunset", result: sunset },
-    ];
-
-    return collection.map((e, i) => {
-      return (
-        <div
-          key={e + " " + i}
-          className="flex items-center justify-between h-8 capitalize daily-details-grid-item"
-        >
-          <span className="text-neutral-100 drop-shadow-md">{e.id}</span>
-          <span className="drop-shadow-md">{e.result}</span>
-        </div>
-      );
-    });
-  };
-
+export const Forecast = ({ data, timezone }: ForecastProps): JSX.Element => {
   return (
     <div className="mt-4">
       <Accordion allowZeroExpanded>
@@ -97,11 +36,11 @@ export const Forecast = (props: ForecastProps): JSX.Element => {
             <AccordionItemHeading>
               <AccordionItemButton>
                 <div className="py-1">
-                  <div className="flex items-center h-16 px-1 py-5 m-1 text-sm bg-transparent border-t border-b border-l border-r border-white hover:border-transparent drop-shadow-md daily-item rounded-xl ">
+                  <div className="flex items-center h-16 px-1 py-5 m-1 text-sm bg-transparent border-t border-b border-l border-r border-white drop-shadow-md rounded-xl hover:border-black hover:font-bold">
                     <span className="flex-auto ml-2 font-semibold day sm:text-lg xl:text-xl 2xl:text-2xl">
                       {dayOfWeek(item)}
                     </span>
-                    <span className="min-max sm:text-lg xl:text-xl 2xl:text-2xl">
+                    <span className="sm:text-lg xl:text-xl 2xl:text-2xl">
                       {toCelsius(item.temp.min)}°C | {toCelsius(item.temp.max)}
                       °C
                       {" | "}
@@ -114,8 +53,8 @@ export const Forecast = (props: ForecastProps): JSX.Element => {
               </AccordionItemButton>
             </AccordionItemHeading>
             <AccordionItemPanel>
-              <div className="grid grid-cols-1 sm:grid-cols-2 flex-auto py-1.5 px-4 gap-y-0 gap-x-5 sm:text-lg md:text-xl 2xl:text-2xl daily-details-grid">
-                <BuildPanel item={item} />
+              <div className="grid grid-cols-1 sm:grid-cols-2 flex-auto py-1.5 px-4 gap-y-0 gap-x-5 sm:text-lg md:text-xl 2xl:text-2xl">
+                <DailyWeatherPanel item={item} timezone={timezone} />
               </div>
             </AccordionItemPanel>
           </AccordionItem>
